server/requirements.txt
http://github.com/mitsuhiko/flask/zipball/master#egg=Flask-dev
gunicorn==19.3.0
Flask-SQLAlchemy
flake8
